+++ carp_context_package/lib/context.dart
@@ -5,6 +5,7 @@
 library context;
 
 import 'dart:async';
+import 'dart:math' as math;
 import 'package:json_annotation/json_annotation.dart';
 import 'package:carp_mobile_sensing/carp_mobile_sensing.dart';
 import 'package:activity_recognition/activity_recognition.dart';