+++ carp_context_package/lib/geofence_measure.dart
@@ -7,8 +7,14 @@
 
 part of context;
 
-/// Eart radius in km.
+/// Earth radius in km.
 const double earthRadius = 6371000.0;
+
+/// Convert degrees to radians.
+double degToRad(num deg) => deg * (math.pi / 180.0);
+
+/// Convert radians to degrees.
+double radToDeg(num rad) => rad * (180.0 / math.pi);
 
 /// Location coordinated in Degrees (i.e. GPS-style).
 @JsonSerializable(fieldRename: FieldRename.snake, includeIfNull: false)
@@ -27,6 +33,21 @@ class Location extends Serializable {
       : latitude = map['latitude'],
         longitude = map['longitude'],
         super();
+
+  /// Returns the approximate distance in meters between this location and the given location.
+  ///
+  /// For distance calculations we use the  'haversine' formula to calculate the great-circle distance between two points.
+  /// See http://www.movable-type.co.uk/scripts/latlong.html for details on how to
+  /// calculate distance, bearing and more between latitude/longitude points.
+  double distanceTo(Location destination) {
+    assert(destination != null);
+    final sDLat = math.sin((degToRad(destination.latitude) - degToRad(latitude)) / 2);
+    final sDLng = math.sin((degToRad(destination.longitude) - degToRad(longitude)) / 2);
+    final a = sDLat * sDLat + sDLng * sDLng * math.cos(degToRad(latitude)) * math.cos(degToRad(destination.latitude));
+    final c = 2 * math.atan2(math.sqrt(a), math.sqrt(1 - a));
+
+    return earthRadius * c;
+  }
 
   static Function get fromJsonFunction => _$LocationFromJson;
   factory Location.fromJson(Map<String, dynamic> json) =>
@@ -50,6 +71,11 @@ class GeofenceMeasure extends Measure {
 
   /// The radius of the geofence in meters.
   double radius;
+
+  /// The dwell time of this geofence in miliseconds.
+  /// If an object is located inside a geofence for more that [dwell] miliseconds,
+  /// [GeofenceState.DWELL] are fired.
+  int dwell;
 
   /// The name of this geofence.
   String name;